pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]

name = "dantalabs"        
version = "0.2.0"          
authors = [
  { name="DantaLabs", email="dev@dantalabs.com" },
]
description = "DantaLabs SDK collection (featuring Maestro)" 
readme = "README.md"
requires-python = ">=3.8"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.24.0,<0.28.0",
    "pydantic>=2.0,<3.0",
    "typer[all]>=0.9.0,<1.0.0",
    "pydantic[email]>=2.0,<=3.0",
    "python-dotenv>=1.1.0 ",
    "PyYAML>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "psycopg>=2.0.0"
]

[project.scripts]

dlm = "dantalabs.cli:app"


[project.urls] 
"Homepage" = "https://github.com/dantalabs/dantalabs-sdk" 
"Bug Tracker" = "https://github.com/dantalabs/dantalabs-sdk/issues"


[tool.setuptools.packages.find]
where = ["."]             # Look for packages in the root directory (where pyproject.toml is)
include = ["dantalabs*"]
